give cloaks their own package
